requirements.txt
numpy>=1.24
scipy>=1.11
Pillow>=10.0
pyyaml>=6.0
tifffile>=2023.1
